move global variables for tests.
