--- detectmatelibrary_tests/__init__.py
@@ -1,3 +1,0 @@
-AUDIT_LOG = "detectmatelibrary_tests/test_folder/audit.log"
-NAMED_TEMPLATES_TXT = "detectmatelibrary_tests/test_folder/test_named_templates.txt"
-NAMED_TEMPLATES_CSV = "detectmatelibrary_tests/test_folder/test_named_templates.csv"
