+++ detectmatelibrary_tests/test_detectors/test_new_event_detector.py
@@ -16,7 +16,7 @@ import detectmatelibrary.schemas as schemas
 from detectmatelibrary.utils.aux import time_test_mode
 from detectmatelibrary.common._core_op._fit_logic import ConfigState, TrainState
 from detectmatelibrary.constants import GLOBAL_EVENT_ID
-from detectmatelibrary_tests import AUDIT_LOG
+from detectmatelibrary_tests.test_pipelines.test_configuration_engine import AUDIT_LOG
 
 
 # Set time test mode for consistent timestamps
